+++ src/rooms/frame_timer.hpp
@@ -5,13 +5,13 @@
 class FrameTimer {
 public:
     explicit FrameTimer(int fps);
-    double delta() const;
+    float delta() const;
     int operator()();
 
 private:
     using Clock = std::chrono::high_resolution_clock;
 
-    const double _delta;
+    const float _delta;
     long long _lastFrame;
     Clock::time_point _startTime;
     Clock::duration _frameDuration;
